+++ resources/views/component.blade.php
@@ -1,0 +1,1 @@
+<livewire:comments :record="$record ?? $this->record" />