+++ src/Infolists/Components/CommentsEntry.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Parallax\FilamentComments\Infolists\Components;
+
+use Filament\Infolists\Components\Entry;
+
+class CommentsEntry extends Entry
+{
+    protected string $view = 'filament-comments::component';
+}