+++ src/Tables/Actions/CommentsAction.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Parallax\FilamentComments\Tables\Actions;
+
+use Filament\Tables\Actions\Action;
+use Filament\Support\Enums\MaxWidth;
+use Illuminate\Contracts\View\View;
+use Illuminate\Database\Eloquent\Model;
+use Parallax\FilamentComments\Models\FilamentComment;
+
+class CommentsAction extends Action
+{
+    public static function getDefaultName(): ?string
+    {
+        return 'comments';
+    }
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this
+            ->icon('heroicon-s-chat-bubble-left-right')
+            ->label(__('filament-comments::filament-comments.comments'))
+            ->color('gray')
+            ->slideOver()
+            ->modalContentFooter(fn (Model $record): View => view('filament-comments::component', [
+                'record' => $record,
+            ]))
+            ->modalHeading(__('filament-comments::filament-comments.modal.heading'))
+            ->modalWidth(MaxWidth::Medium)
+            ->modalSubmitAction(false)
+            ->modalCancelAction(false)
+            ->visible(fn (): bool => auth()->user()->can('create', FilamentComment::class));
+    }
+}