+++ src/Testing/TestsFilamentComments.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Parallax\FilamentComments\Testing;
+
+use Livewire\Features\SupportTesting\Testable;
+
+/**
+ * @mixin Testable
+ */
+class TestsFilamentComments
+{
+    //
+}